// src/api-client/models/cancellation-reason-code.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Cancellation reason as a fixed code.
 * @export
 * @enum {string}
 */

export const CancellationReasonCode = {
    CannotMakeWindow: 'CANNOT_MAKE_WINDOW',
    VehicleBreakdown: 'VEHICLE_BREAKDOWN',
    VehicleCapacityReached: 'VEHICLE_CAPACITY_REACHED',
    StorageCapacityReached: 'STORAGE_CAPACITY_REACHED',
    StoreLogisticsDelay: 'STORE_LOGISTICS_DELAY',
    NoLongerNeeded: 'NO_LONGER_NEEDED',
    DriverNoShow: 'DRIVER_NO_SHOW',
    Other: 'OTHER'
} as const;

export type CancellationReasonCode = typeof CancellationReasonCode[keyof typeof CancellationReasonCode];