// src/api-client/models/reservation-window.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Which bucket the reservation belongs in.
 * @export
 * @enum {string}
 */

export const ReservationWindow = {
    Today: 'today',
    Upcoming: 'upcoming',
    Past: 'past'
} as const;

export type ReservationWindow = typeof ReservationWindow[keyof typeof ReservationWindow];